// src/core/schema.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};
// ...
export type Hit = {
  __typename?: 'Hit';
  createdAt: Scalars['DateTime'];
  hashId: Scalars['ID'];
  id: Scalars['Int'];
  updatedAt?: Maybe<Scalars['DateTime']>;
};

export type HitInput = {
  postHashId: Scalars['String'];
};

export type LikeInput = {
  postHashId: Scalars['String'];
};

export type Likes = {
  __typename?: 'Likes';
  createdAt: Scalars['DateTime'];
  hashId: Scalars['ID'];
  id: Scalars['Int'];
  updatedAt?: Maybe<Scalars['DateTime']>;
};
// ...
export type PageEdge = {
  __typename?: 'PageEdge';
  cursor: Scalars['Float'];
  node: Post;
};
// ...
export type Post = {
  __typename?: 'Post';
  content?: Maybe<Scalars['String']>;
  createdAt: Scalars['DateTime'];
  desc: Scalars['String'];
  hashId: Scalars['ID'];
  hit?: Maybe<Array<Hit>>;
  id: Scalars['Int'];
  isPublished: Scalars['Boolean'];
  likes?: Maybe<Array<Likes>>;
  open: Scalars['Boolean'];
  originalContent?: Maybe<Scalars['String']>;
  reply?: Maybe<Array<Reply>>;
  tag: Array<Tag>;
  thumbnail?: Maybe<Scalars['String']>;
  title: Scalars['String'];
  updatedAt?: Maybe<Scalars['DateTime']>;
};
// ...
export type Reply = {
  __typename?: 'Reply';
  bgroup: Scalars['Float'];
  comment: Scalars['String'];
  createdAt: Scalars['DateTime'];
  depth: Scalars['Float'];
  hashId: Scalars['ID'];
  id: Scalars['Float'];
  parentId?: Maybe<Scalars['Float']>;
  postId?: Maybe<Scalars['Float']>;
  sorts: Scalars['Float'];
  updatedAt?: Maybe<Scalars['DateTime']>;
  writer: Scalars['String'];
};
// ...
export type Tag = {
  __typename?: 'Tag';
  createdAt: Scalars['DateTime'];
  hashId: Scalars['ID'];
  id: Scalars['Int'];
  post: Array<Post>;
  tag: Scalars['String'];
  updatedAt?: Maybe<Scalars['DateTime']>;
};